running evaluate at scale

+++ python/evaluate.py
@@ -6,7 +6,7 @@ import json
 import os
 import logging
 import math
-import numpy
+#import numpy
 import re
 import shutil
 import sys
@@ -24,8 +24,8 @@ from chemotext_util import SerializationUtil as SUtil
 from chemotext_util import SparkUtil
 from pyspark.sql import SQLContext
 
-from pyspark.mllib.classification import LogisticRegressionWithLBFGS, LogisticRegressionModel
-from pyspark.mllib.regression import LabeledPoint
+#from pyspark.mllib.classification import LogisticRegressionWithLBFGS, LogisticRegressionModel
+#from pyspark.mllib.regression import LabeledPoint
 
 logger = LoggingUtil.init_logging (__file__)
 
@@ -190,13 +190,18 @@ def expand_ref_binaries (binary):
     return result
 def get_facts (sc, conf):
     sqlContext = SQLContext(sc)
-    reference_binaries = [
-        load_facts (sqlContext, conf.ctdAB, 0, 3, 10),
-        load_facts (sqlContext, conf.ctdBC, 0, 2, 8),
-        load_facts (sqlContext, conf.ctdAC, 0, 3, 9)
-    ]
+    ab = load_facts (sqlContext, conf.ctdAB, 0, 3, 10)
+    bc = load_facts (sqlContext, conf.ctdBC, 0, 2, 8)
+    ac = load_facts (sqlContext, conf.ctdAC, 0, 3, 9)
+    reference_binaries = [ ab, bc, ac ]
+    ab.unpersist ()
+    bc.unpersist ()
+    ac.unpersist ()
+    ab = None
+    bc = None
+    ac = None
     return sc.union (reference_binaries). \
-        flatMap (expand_ref_binaries)
+        flatMap (expand_ref_binaries).cache ()
 
 def get_article_guesses (article_path):
     logger = LoggingUtil.init_logging (__file__)
@@ -208,11 +213,26 @@ def distance (b):
     return b
 def get_guesses (sc, conf):
     articles = glob.glob (os.path.join (conf.input_dir, "*fxml.json"))
-    articles = sc.parallelize (articles)
+    articles = sc.parallelize (articles) #.sample (False, 0.5, 81)
     return articles.\
         flatMap (lambda article : get_article_guesses (article)). \
         map (lambda b : ( make_key (b.L, b.R, b.pmid), distance (b) ) ). \
         reduceByKey (lambda x, y : x if x.dist < y.dist else x)
+
+def get_guesses (sc, conf, slices, slice_n):
+    articles = glob.glob (os.path.join (conf.input_dir, "*fxml.json"))
+    slice_size = int (len (articles) / slices)
+    offset = slice_size * slice_n
+    print "--------------------slice size {0}".format (slice_size)
+    print "--------------------    offset {0}".format (offset)
+    articles = articles [ offset : offset + slice_size ]
+    articles = sc.parallelize (articles, 20) #.sample (False, 0.5, 81)
+    skiplist = [ 'for', 'was' ]
+    return articles.\
+        flatMap (lambda article : get_article_guesses (article)). \
+        filter (lambda b : b.L not in skiplist and b.R not in skiplist ). \
+        map (lambda b : ( make_key (b.L, b.R, b.pmid), distance (b) ) ). \
+        reduceByKey (lambda x, y : x if x.dist < y.dist else x).cache ()
 
 def mark_binary (binary, fact=True):
     binary.fact = fact
@@ -239,26 +259,53 @@ def annotate (guesses, facts):
     false_negative = facts.subtractByKey (true_positive)
     trace_set (trace_level, "FNeg", false_negative)
 
-    return true_positive.\
-        union (false_positive).\
-        union (false_negative).\
-        map (lambda v: v[1])
+    union = true_positive.\
+            union (false_positive).\
+            union (false_negative)
+    true_positive.unpersist ()
+    false_positive.unpersist ()
+    false_negative.unpersist ()
+    true_positive = None
+    false_positive = None
+    false_negative = None
+    return union.map (lambda v: v[1])
 
 def evaluate_articles (conf):
+    
     logger.info ("Evaluating Chemotext2 output: {0}".format (conf.input_dir))
     sc = SparkUtil.get_spark_context (conf)
 
+    logger.info ("Loading facts")
     facts = get_facts (sc, conf)
-    guesses = get_guesses (sc, conf)
-    annotated = annotate (guesses, facts)
+    logger.info ("Loading guesses")
 
-    shutil.rmtree ("annotated")
-    shutil.rmtree ("machine")
+    '''
+    guesses = get_guesses (sc, conf)
+    logger.info ("Annotating")
+    annotated = annotate (guesses, facts)
+    if os.path.exists (conf.output_dir):
+        shutil.rmtree (conf.output_dir)
+    if os.path.exists ("machine"):
+        shutil.rmtree ("machine")
+    output_path = os.path.join (conf.output_dir, "annotated")
     annotated.\
         map(lambda b : json.dumps (b, cls=BinaryEncoder)). \
-        saveAsTextFile ("annotated")
+        saveAsTextFile ("file://" + output_path)
+    '''
+    slices = 10000
+    for slice_n in range (0, slices):
+        output_dir = os.path.join (conf.output_dir, "annotated", str(slice_n))
+        if not os.path.exists (output_dir):
+            guesses = get_guesses (sc, conf, slices, slice_n)
+            annotated = annotate (guesses, facts)
+            logger.info ("Generating annotated output for " + output_dir)
+            annotated.\
+                map(lambda b : json.dumps (b, cls=BinaryEncoder)). \
+                saveAsTextFile ("file://" + output_dir)
+#        for x in range (2, 7):
+#            subprocess.call ([ "ssh", "stars-c{0}.edc.renci.org".format (x), "rm -rf /tmp/mesos/slaves/*" ])
 
-    train_log_reg (sc, annotated)
+#    train_log_reg (sc, annotated)
 
 def train_log_reg (sc, annotated):
     def label (b):
@@ -288,18 +335,20 @@ def main ():
     Parse command line arguments for the evaluation pipeline.
     '''
     parser = argparse.ArgumentParser()
-    parser.add_argument("--host",  help="Mesos master host")
-    parser.add_argument("--name",  help="Spark framework name")
-    parser.add_argument("--input", help="Output directory for a Chemotext2 run.")
-    parser.add_argument("--venv",  help="Path to Python virtual environment to use")
-    parser.add_argument("--ctdAB", help="Path to CTD AB data")
-    parser.add_argument("--ctdBC", help="Path to CTD BC data")
-    parser.add_argument("--ctdAC", help="Path to CTD AC data")
+    parser.add_argument("--host",   help="Mesos master host")
+    parser.add_argument("--name",   help="Spark framework name")
+    parser.add_argument("--input",  help="Output directory for a Chemotext2 run.")
+    parser.add_argument("--output", help="Output directory for evaluation.")
+    parser.add_argument("--venv",   help="Path to Python virtual environment to use")
+    parser.add_argument("--ctdAB",  help="Path to CTD AB data")
+    parser.add_argument("--ctdBC",  help="Path to CTD BC data")
+    parser.add_argument("--ctdAC",  help="Path to CTD AC data")
     args = parser.parse_args()
     conf = EvaluateConf (host           = args.host,
                          venv           = args.venv,
                          framework_name = args.name,
                          input_dir      = args.input,
+                         output_dir     = args.output,
                          ctdAB          = args.ctdAB,
                          ctdBC          = args.ctdBC,
                          ctdAC          = args.ctdAC)
